+++ RiotSharp/CurrentGameEndpoint/Enums/GameQueueType.cs
@@ -1,6 +1,7 @@
 ﻿using Newtonsoft.Json;
+using RiotSharp.CurrentGameEndpoint.Enums.Converters;
 
-namespace RiotSharp.CurrentGameEndpoint.Converters
+namespace RiotSharp.CurrentGameEndpoint.Enums
 {
     /// <summary>
     /// queueType and gameQueueConfigId subtype of CurrentGame API
